requirements.txt
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
boto3==1.28.0
botocore==1.31.85
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
commonmark==0.9.1
Deprecated==1.2.18
distlib==0.3.9
dnspython==2.7.0
email-validator==2.3.0
eventlet==0.40.2
filelock==3.18.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Limiter==3.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
instaloader==4.14.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
limits==5.6.0
MarkupSafe==3.0.2
ordered-set==4.1.0
packaging==25.0
passlib==1.7.4
Pillow==10.0.0
platformdirs==4.3.7
psutil==7.1.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.2
python-socketio==5.13.0
razorpay==1.4.2
requests==2.31.0
rich==12.6.0
s3transfer==0.6.2
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.31
typing_extensions==4.15.0
urllib3==2.0.7
virtualenv==20.30.0
Werkzeug==2.3.7
wheel==0.45.1
wrapt==1.17.3
wsproto==1.2.0
WTForms==3.2.1
